+++ docs/changes.md
@@ -9,6 +9,8 @@
     * the **state** attribute was removed from the index file
     * replaced option **--have-php-index** by the option **--topic-format *&lt;FORMAT&gt;*** which directly defines how each of a blog entries topics shall be rendered when embedding it into the template
     * removed options **--default-author *&lt;NAME&gt;***, **--default-copyright-date *&lt;DATE&gt;***, **--default-state *&lt;STATE&gt;*** and introduced replacements with defaults instead
+* new
+    * the identation level of the index file can now be set using the option **--index-indent *&lt;INT&gt;***
 
 
 ## gresiblos-0.4.2 (29.12.2024)